add confirm before permanent deletion; add to last update

+++ frontend/js/public/deleteButton.js
@@ -1,8 +1,10 @@
 $(".delete-button").each((index, button) => {
   var id = $(button).data("id");
   $(button).on("click", async () => {
-    var response = await fetch(`/delete/${id}`, { method: "DELETE" });
-    console.log(response);
-    if (response.ok) $(`#${id}`).remove();
+    if (confirm("Are you sure? This cannot be undone.")) {
+      var response = await fetch(`/delete/${id}`, { method: "DELETE" });
+      console.log(response);
+      if (response.ok) $(`#${id}`).remove();
+    }
   });
 });
